change: moved constants to respective files
