+++ src/models/transaction.model.ts
@@ -11,3 +11,30 @@ export interface TokenTransaction {
   totalValue: number
   txnHash: string
 }
+
+export interface TransactionObj {
+  from: string
+  to: string
+  timestamp: number
+  hash: string
+  blockchain: string
+  value: string
+  token: string | null
+  tokenIcon: string | null
+  gas: string
+  type: string
+  status: string
+  isVerified: boolean
+  tokenAddress: string
+  additionalData: {
+    recipient?: string
+    amount?: string
+    fromTokenAddress?: string
+    fromToken?: string
+    fromTokenIcon?: string
+    fromTokenValue?: string
+    toTokenAddress?: string
+    toToken?: string
+    toTokenIcon?: string
+  }
+}
